CI: remove ccache
Remove it for now. Its cache is handled automatically, but only when the language is set to C/C++.
In that mode however there are issues with build configuration, hence it had to be generic and with manual handling
of the ccache cache folder (giving the path to cache folder by hand).

Signed-off-by: Szczepan Zalega <szczepan@nitrokey.com>