Fix logical negation operator
CMake expresses logical negation as `NOT` and not as `!`.
See: https://cmake.org/cmake/help/latest/command/if.html#condition-syntax

The present syntax causes the system libnitrokey to be always ignored,
and the bundled libnitrokey to be always used. This defeats the purpose
of the check for a system-provided libnitrokey.

Signed-off-by: Nicolas Benes <nbenes.gh@xandea.de>